feat: Implement Vocabulary Engine (Service, Models, Schemas) (#8) (#9)
feat: Implement Vocabulary Engine (Service, Models, Schemas)

Establishes the core engine for Concept search and retrieval, ensuring compatibility with legacy WebAPI logic while introducing modern caching and strict typing.

**Vocabulary Service & Logic:**
- Implemented `VocabularyService` with robust concept search, filtering (Domain, Vocabulary, Class), and pagination.
- Integrated Redis caching for concept lookups using Pydantic deserialization.
- Added support for legacy WebAPI logic mappings (e.g., handling Measurement domains, standard concept `N->Null` transformations).

**Data Models & Schemas:**
- Created Read-Only SQLAlchemy models: `Concept`, `Vocabulary`, `Domain`, `ConceptClass`.
- Defined Pydantic schemas with aliasing for full API compatibility (handling uppercase request keys vs. camelCase response keys).

**Dependencies & Config:**
- Added `pytest-mock` to project dependencies and updated `poetry.lock`.
- Configured strict typing overrides to resolve runtime TypeErrors with Redis generics while satisfying Mypy.

**Quality Assurance:**
- Added comprehensive tests achieving 100% coverage for all new components.
- Removed conflicting legacy tests (`tests/services/test_vocabulary.py`).
- Resolved all Ruff formatting, Mypy strict type errors, and added missing license headers.

https://github.com/CoReason-AI/omop-atlas-backend/commit/56509344120805abaab3c9e2525a72e35e140701

+++ src/omop_atlas_backend/models/vocabulary.py
@@ -11,7 +11,7 @@
 from datetime import date
 from typing import Optional
 
-from sqlalchemy import Date, Integer, String
+from sqlalchemy import BigInteger, Date, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from omop_atlas_backend.models.base import Base
@@ -19,9 +19,9 @@ from omop_atlas_backend.models.base import Base
 
 class Concept(Base):
     __tablename__ = "concept"
-    # __table_args__ = {"schema": "cdm"} # To be configured
 
-    concept_id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    # Read-only model
+    concept_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     concept_name: Mapped[str] = mapped_column(String(255))
     domain_id: Mapped[str] = mapped_column(String(20))
     vocabulary_id: Mapped[str] = mapped_column(String(20))
